+++ crates/bitcoin-precompiles/src/precompiles/mod.rs
@@ -1,0 +1,4 @@
+pub mod send_raw_transaction;
+pub mod get_block_count;
+pub mod decode_raw_transaction;
+pub mod check_signature;
